test: reduce circular dependencies in tests
Separate test helpers from integration test helpers,
so we can run most tests without loading the cssnano package,
which causes loading cssnano-preset-default and so most of the repository.

I noticed this issue because pointing the `main` field to the wrong file in the
package.json of postcss-normalize-timing-functions
causes every test to fail even it does not depend on postcss-normalize-timing-functions.
