+++ packages/cssnano-preset-default/src/__tests__/css-declaration-sorter.js
@@ -1,5 +1,5 @@
 import { test } from 'uvu';
-import { processCSSWithPresetFactory } from '../../../../util/testHelpers';
+import { processCSSWithPresetFactory } from '../../../../util/integrationTestHelpers';
 import preset from '..';
 
 const { processCSS, passthroughCSS } = processCSSWithPresetFactory(preset);
